Add files via upload
The attached product (A) displays a detailed map of Red Mountain Park, a 1,500 acre recreational preserve located in Birmingham, Alabama. I was tasked with providing an updated park map that incorporates GIS data.

Imagery from an older park brochure was used to geo-reference the terrain to a satellite base map. This was achieved by applying control points to fixed map features. I used the WGS 1984 Web Mercator coordinate system to obtain the most accurate geographic measurements available for a two-dimensional layout. 

Trails, points of interest, and the park perimeter were given customized colors and are represented by a series of distinct symbols on a simple-to-read legend. Each feature on the legend is also represented in the geodatabase as its own feature class (B) and can be customized per the client’s preferences.